Move settings.LOGGING to custom_logging
